CommonElements.vue作成、vue-router実装
